simplify and slim the footer
